Clean up project structure for modules relating to `phases run`
